requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2

# Frontend Dependencies
streamlit==1.28.1
requests==2.31.0

# Gemini API
google-generativeai==0.3.2

# Utilities
python-multipart==0.0.6

